python-itom-task/requirements.txt
ansible==2.3.1.0
docker==3.5.0
requests==2.13.0
redis==2.10.6
PyJWT==1.6.1